+++ src/physics/thermalnoise.cu
@@ -58,13 +58,20 @@ Field evalThermalNoise(const Ferromagnet* magnet) {
   real mean = 0.0;
   real stddev = 1.0;
 
-  real* tmp;
-  cudaMalloc(&tmp, (N + (N % 2)) * sizeof(real));
-  for (int c = 0; c < 3; c++) {
-    generateRandNormal(magnet->randomGenerator, tmp, N + (N % 2), mean, stddev);
-    cudaMemcpy(noise.device_ptr(c), tmp, N * sizeof(real), cudaMemcpyDeviceToDevice);
+  // CUDA RNG is only compatible with even N.
+  if (N % 2) {
+    real* tmp;
+    cudaMalloc(&tmp, (N + (N % 2)) * sizeof(real));
+    for (int c = 0; c < 3; c++) {
+      generateRandNormal(magnet->randomGenerator, tmp, N + (N % 2), mean, stddev);
+      cudaMemcpy(noise.device_ptr(c), tmp, N * sizeof(real), cudaMemcpyDeviceToDevice);
+    }
+    cudaFree(tmp);
   }
-  cudaFree(tmp);
+  else {
+    for (int c = 0; c < 3; c++)
+      generateRandNormal(magnet->randomGenerator, noise.device_ptr(c), N, mean, stddev);
+  }
 
   auto msat = magnet->msat.cu();
   auto alpha = magnet->alpha.cu();
